Day10-FinalHackathon/python-backend-for-ai/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.0.350
langchain-google-genai==0.0.8
langchain-community==0.0.13
google-generativeai==0.3.2
# faiss-cpu==1.7.4
pydantic<2.0
python-dotenv==1.0.0
chromadb==0.4.18